Move PFAR diversification to its own file
