+++ src/poprox_recommender/default.py
@@ -1,4 +1,3 @@
-import math
 import random
 import sys
 from dataclasses import dataclass
@@ -15,6 +14,7 @@ from transformers import AutoTokenizer
 
 from poprox_concepts import Article, ClickHistory, InterestProfile
 from poprox_recommender.diversifiers.mmr import mmr_diversification
+from poprox_recommender.diversifiers.pfar import pfar_diversification
 from poprox_recommender.model.nrms import NRMS
 from poprox_recommender.paths import model_file_path
 from poprox_recommender.topics import normalized_topic_count, user_topic_preference
@@ -113,60 +113,6 @@ def build_user_embedding(click_history: ClickHistory, article_embeddings, model,
     )
 
     return model.get_user_vector(clicked_news_vector)
-
-
-def pfar_diversification(relevance_scores, articles, topic_preferences, lamb, tau, topk):
-    # p(v|u) + lamb*tau \sum_{d \in D} P(d|u)I{v \in d} \prod_{i \in S} I{i \in d} for each user
-
-    if tau is None:
-        tau = 0
-        for topic, weight in topic_preferences.items():
-            if weight > 0:
-                tau -= weight * math.log(weight)
-    else:
-        tau = float(tau)
-
-    S = []  # final recommendation LIST[candidate index]
-    initial_item = relevance_scores.argmax()
-    S.append(initial_item)
-
-    S_topic = set()
-    article = articles[int(initial_item)]
-    S_topic.update(extract_general_topics(article))
-
-    for k in range(topk - 1):
-        candidate_idx = None
-        best_score = float("-inf")
-
-        for i, relevance_i in enumerate(relevance_scores):  # iterate R for next item
-            if i in S:
-                continue
-            product = 1
-            summation = 0
-
-            candidate_topics = set(extract_general_topics(articles[int(i)]))
-
-            for topic in candidate_topics:
-                if topic in S_topic:
-                    product = 0
-                    break
-
-            for topic in candidate_topics:
-                if topic in topic_preferences:
-                    summation += topic_preferences[topic]
-
-            pfar_score_i = relevance_i + lamb * tau * summation * product
-
-            if pfar_score_i > best_score:
-                best_score = pfar_score_i
-                candidate_idx = i
-
-        if candidate_idx is not None:
-            candidate_topics = set(extract_general_topics(articles[int(candidate_idx)]))
-            S.append(candidate_idx)
-            S_topic.update(candidate_topics)
-
-    return S  # LIST(candidate index)
 
 
 def generate_recommendations(
